Fix few requests, add comments to back-end

+++ web-game-chess-backend/chess.Application/Requests/GameRequests/AbortSearch/AbortSearchRequestHandler.cs
@@ -1,5 +1,6 @@
 ﻿
 using chess.Application.Repositories;
+using chess.Application.Services;
 using chess.Shared.Exceptions;
 using MediatR;
 
@@ -8,18 +9,25 @@ namespace chess.Application.Requests.GameRequests.AbortSearch;
 public class AbortSearchRequestHandler : IRequestHandler<AbortSearchRequest> {
 
     private readonly IPlayerRepository _playerRepository;
+    private readonly IUserContextService _userContextService;
 
-    public AbortSearchRequestHandler(IPlayerRepository playerRepository) {
+    public AbortSearchRequestHandler(IPlayerRepository playerRepository, IUserContextService userContextService) {
         _playerRepository = playerRepository;
+        _userContextService = userContextService;
     }
 
     public async Task Handle(AbortSearchRequest request, CancellationToken cancellationToken) {
 
+        var userId = _userContextService.GetUserId();
+
         var playerToDelete = await _playerRepository.GetById(request.PlayerId) 
             ?? throw new NotFoundException("Player not found.");
 
+        if (userId != playerToDelete.UserId)
+            throw new BadRequestException("Player do not belong to user.");
+
         if (playerToDelete.IsPlaying)
-            throw new BadRequestException("Can not remove player.");
+            throw new BadRequestException("Can not remove player that is in game.");
 
         await _playerRepository.Delete(playerToDelete);
     }
